+++ src/erhe/geometry/erhe_geometry/shapes/sphere.hpp
@@ -4,6 +4,6 @@ namespace GEO { class Mesh; }
 
 namespace erhe::geometry::shapes {
 
-void make_sphere(GEO::Mesh& mesh, double radius, unsigned int slice_count, unsigned int stack_division);
+void make_sphere(GEO::Mesh& mesh, float radius, unsigned int slice_count, unsigned int stack_division);
 
 } // namespace erhe::geometry::shapes
