+++ src/erhe/geometry/erhe_geometry/shapes/torus.hpp
@@ -4,6 +4,6 @@ namespace GEO { class Mesh; };
 
 namespace erhe::geometry::shapes {
 
-void make_torus(GEO::Mesh& mesh, double major_radius, double minor_radius, int major_axis_steps, int minor_axis_steps);
+void make_torus(GEO::Mesh& mesh, float major_radius, float minor_radius, int major_axis_steps, int minor_axis_steps);
 
 } // namespace erhe::geometry::shapes
